Revert "Optimize string parsing" (#27)
This reverts commit 0ff4ecec0aaf822f650e704e4ed67c2b6a01a6c6.

This fixes a seg-fault that was introduced with 0.1.14. The reason for that segfault is that it introduced an issue of us reading after the input data od possibly crossing page boundaries.

Better handling of relocated data might solve that but rust could get into the way here. Worth an investigation later.