Fixup benchmarks and dependencies
The benches now include simd_json::to_tape, drop the C++ FFI crate
because it's not on cargo and anyway out of date, and use thinLTO
to be more representative.

This reduces the dependencies needed for cargo test from 133
to 85 by putting benchmarking-only dependencies behind a feature.
It's ~30% faster to compile.